Reset `before` value to None in _query_txids between addresses for report_sol (#6)
In cases that `max_queries` is less than the total number of queries possible for an address, the `before` value would return the value needed to make the next call for that address rather than None. In that case, not resetting the value of before to None when moving onto the next wallet could cause issues with the `RpcAPI.get_txids` call for the next wallet.