Cleanup to report_osmo get_txs function (#7)
* Reset `before` value to None in _query_txids between addresses for report_sol

In cases that `max_queries` is less than the total number of queries possible for an address, the `before` value would return the value needed to make the next call for that address rather than None. In that case, not resetting the value of before to None when moving onto the next wallet could cause issues with the `RpcAPI.get_txids` call for the next wallet.

* Cleanup osmo _get_txs function to use a calculated offset

Rather than explicitly keep track of the offest, calculate the value based on index var and LIMIT.

Add comment to explain the break condition.